Allow 'W'ear and 'T'ake off to work on rings/amulets/blindfolds (wear_unified)
Wear and Take off commands can now work on rings, amulets and
blindfolds i.e. they appear in the object selection prompts for these
commands, and are put on or removed respectively, allowing 'W' and 'T'
to substitute for 'P'ut on and 'R'emove.

This behavior can be toggled with the new 'wear_unified' boolean option
(default on).

Preserves save compatibility.